rebuild_pipeline/requirements.txt
boto3>=1.35.0
pymongo>=4.8.0
python-dotenv>=1.0.0
